+++ packages/layer-composer/src/generators/heatmap/heatmap.ts
@@ -119,6 +119,28 @@ class HeatmapGenerator {
           group: config.metadata?.group || Group.Heatmap,
         },
       },
+      {
+        id: `${config.id}_interaction`,
+        source: config.id,
+        'source-layer': 'temporalgrid',
+        type: 'line',
+        paint: {
+          'line-color': 'white',
+          'line-width': [
+            'case',
+            ['boolean', ['feature-state', 'hover'], false],
+            4,
+            ['boolean', ['feature-state', 'click'], false],
+            4,
+            0,
+          ],
+          'line-offset': -2,
+        },
+        metadata: {
+          interactive: false,
+          group: Group.Heatmap,
+        },
+      },
     ]
   }